\name{toy.smps}
\docType{data}
\alias{toy.smps}
\title{
Simulated data set containing 17 interval-valued observations of 2 variables.
}
\description{
This data set was used for the illustrations in A. Wiencierz, M. Cattaneo (2012).
}
\usage{data('toy.smps')}
\format{A \code{data.frame} containing 17 rows and 4 columns, i.e. 2 neighboring columns for each variable, the first column for the left endpoints of the interval observations, the second for the right endpoints.}
\references{
M. Cattaneo, A. Wiencierz (2012b). On the implementation of LIR: the case of simple linear regression with interval data. Technical Report No. 127. Department of Statistics. LMU Munich.

A. Wiencierz, M. Cattaneo (2012). An exact algorithm for Likelihood-based Imprecise Regression in the case of simple linear regression with interval data. (Accepted for the 6th International Conference on Soft Methods in Probability and Statistics (SMPS 2012). Publication in the series Advances in Intelligent and Soft Computing. Springer-Verlag.)

M. Cattaneo, A. Wiencierz (2012a). Likelihood-based Imprecise Regression. International Journal of Approximate Reasoning. In Press. (Uncorrected Proof: \url{http://www.sciencedirect.com/science/article/pii/S0888613X12000862})
}
